+++ sdk/identity/identity/src/credentials/managedIdentityCredential/arcMsi.ts
@@ -8,19 +8,24 @@ import {
 } from "@azure/core-rest-pipeline";
 import { AccessToken, GetTokenOptions } from "@azure/core-auth";
 import { readFile } from "fs";
-import { MSI } from "./models";
+import { MSI, MSIConfiguration } from "./models";
 import { credentialLogger } from "../../util/logging";
 import { IdentityClient } from "../../client/identityClient";
-import { msiGenericGetToken } from "./utils";
+import { mapScopesToResource, msiGenericGetToken } from "./utils";
 import { azureArcAPIVersion } from "./constants";
 import { AuthenticationError } from "../../client/errors";
 
-const logger = credentialLogger("ManagedIdentityCredential - ArcMSI");
+const msiName = "ManagedIdentityCredential - Azure Arc MSI";
+const logger = credentialLogger(msiName);
 
 // Azure Arc MSI doesn't have a special expiresIn parser.
 const expiresInParser = undefined;
 
-function prepareRequestOptions(resource?: string): PipelineRequestOptions {
+function prepareRequestOptions(scopes: string | string[]): PipelineRequestOptions {
+  const resource = mapScopesToResource(scopes);
+  if (!resource) {
+    throw new Error(`${msiName}: Multiple scopes are not supported.`);
+  }
   const queryParameters: any = {
     resource,
     "api-version": azureArcAPIVersion
@@ -30,7 +35,7 @@ function prepareRequestOptions(resource?: string): PipelineRequestOptions {
 
   // This error should not bubble up, since we verify that this environment variable is defined in the isAvailable() method defined below.
   if (!process.env.IDENTITY_ENDPOINT) {
-    throw new Error("Missing environment variable: IDENTITY_ENDPOINT");
+    throw new Error(`${msiName}: Missing environment variable: IDENTITY_ENDPOINT`);
   }
 
   return createPipelineRequest({
@@ -69,7 +74,7 @@ async function filePathRequest(
     }
     throw new AuthenticationError(
       response.status,
-      `To authenticate with Azure Arc MSI, status code 401 is expected on the first request.${message}`
+      `${msiName}: To authenticate with Azure Arc MSI, status code 401 is expected on the first request. ${message}`
     );
   }
 
@@ -82,24 +87,31 @@ async function filePathRequest(
 }
 
 export const arcMsi: MSI = {
-  async isAvailable(): Promise<boolean> {
+  async isAvailable(scopes): Promise<boolean> {
+    const resource = mapScopesToResource(scopes);
+    if (!resource) {
+      logger.info(`${msiName}: Unavailable. Multiple scopes are not supported.`);
+      return false;
+    }
     const result = Boolean(process.env.IMDS_ENDPOINT && process.env.IDENTITY_ENDPOINT);
     if (!result) {
-      logger.info("The Azure Arc MSI is unavailable.");
+      logger.info(
+        `${msiName}: The environment variables needed are: IMDS_ENDPOINT and IDENTITY_ENDPOINT`
+      );
     }
     return result;
   },
   async getToken(
-    identityClient: IdentityClient,
-    resource?: string,
-    clientId?: string,
+    configuration: MSIConfiguration,
     getTokenOptions: GetTokenOptions = {}
   ): Promise<AccessToken | null> {
-    logger.info(`Using the Azure Arc MSI to authenticate.`);
+    const { identityClient, scopes, clientId } = configuration;
+
+    logger.info(`${msiName}: Authenticating.`);
 
     if (clientId) {
       throw new Error(
-        "User assigned identity is not supported by the Azure Arc Managed Identity Endpoint. To authenticate with the system assigned identity omit the client id when constructing the ManagedIdentityCredential, or if authenticating with the DefaultAzureCredential ensure the AZURE_CLIENT_ID environment variable is not set."
+        `${msiName}: User assigned identity is not supported by the Azure Arc Managed Identity Endpoint. To authenticate with the system assigned identity, omit the client id when constructing the ManagedIdentityCredential, or if authenticating with the DefaultAzureCredential ensure the AZURE_CLIENT_ID environment variable is not set.`
       );
     }
 
@@ -107,14 +119,14 @@ export const arcMsi: MSI = {
       disableJsonStringifyOnBody: true,
       deserializationMapper: undefined,
       abortSignal: getTokenOptions.abortSignal,
-      ...prepareRequestOptions(resource),
+      ...prepareRequestOptions(scopes),
       allowInsecureConnection: true
     };
 
     const filePath = await filePathRequest(identityClient, requestOptions);
 
     if (!filePath) {
-      throw new Error("Azure Arc MSI failed to find the token file.");
+      throw new Error(`${msiName}: Failed to find the token file.`);
     }
 
     const key = await readFileAsync(filePath, { encoding: "utf-8" });