+++ sdk/identity/identity/src/index.ts
@@ -59,6 +59,13 @@ export {
   VisualStudioCodeCredentialOptions
 } from "./credentials/visualStudioCodeCredential";
 
+export {
+  OnBehalfOfCredential,
+  OnBehalfOfCredentialSecretConfiguration,
+  OnBehalfOfCredentialCertificateConfiguration
+} from "./credentials/onBehalfOfCredential";
+export { OnBehalfOfCredentialOptions } from "./credentials/onBehalfOfCredentialOptions";
+
 export { TokenCachePersistenceOptions } from "./msal/nodeFlows/tokenCachePersistenceOptions";
 
 export {