+++ sdk/network/arm-network/src/operations/vpnSitesConfiguration.ts
@@ -11,15 +11,14 @@ import * as coreClient from "@azure/core-client";
 import * as Mappers from "../models/mappers";
 import * as Parameters from "../models/parameters";
 import { NetworkManagementClientContext } from "../networkManagementClientContext";
-import { PollerLike, PollOperationState } from "@azure/core-lro";
-import { LroEngine } from "../lro";
-import { CoreClientLro, shouldDeserializeLro } from "../coreClientLro";
+import { PollerLike, PollOperationState, LroEngine } from "@azure/core-lro";
+import { LroImpl } from "../lroImpl";
 import {
   GetVpnSitesConfigurationRequest,
   VpnSitesConfigurationDownloadOptionalParams
 } from "../models";
 
-/** Class representing a VpnSitesConfiguration. */
+/** Class containing VpnSitesConfiguration operations. */
 export class VpnSitesConfigurationImpl implements VpnSitesConfiguration {
   private readonly client: NetworkManagementClientContext;
 
@@ -83,13 +82,16 @@ export class VpnSitesConfigurationImpl implements VpnSitesConfiguration {
       };
     };
 
-    const lro = new CoreClientLro(
+    const lro = new LroImpl(
       sendOperation,
       { resourceGroupName, virtualWANName, request, options },
-      downloadOperationSpec,
-      "location"
+      downloadOperationSpec
     );
-    return new LroEngine(lro, { intervalInMs: options?.updateIntervalInMs });
+    return new LroEngine(lro, {
+      resumeFrom: options?.resumeFrom,
+      intervalInMs: options?.updateIntervalInMs,
+      lroResourceLocationConfig: "location"
+    });
   }
 
   /**