+++ sdk/tables/data-tables/recordings/node/tableclient_sasconnectionstring_createentity_getentity_and_delete/recording_should_createentity_with_int64.js
@@ -1,6 +1,6 @@
 let nock = require('nock');
 
-module.exports.hash = "bddae002daeb34fa47424e45c0152faf";
+module.exports.hash = "9328941615eb80fa4e895157041f336e";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
@@ -13,15 +13,15 @@ nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":tru
   'Content-Length',
   '0',
   'ETag',
-  `W/"datetime'2021-06-09T16%3A21%3A45.3613343Z'"`,
+  `W/"datetime'2021-09-07T16%3A24%3A07.3017257Z'"`,
   'Location',
   "https://fakeaccount.table.core.windows.net/tableClientTestSASConnectionStringnode(PartitionKey='P4_node',RowKey='R4')",
   'Server',
   'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '654ef0c6-9002-001e-724b-5d5b52000000',
+  '007653ef-e002-0038-2604-a4c0e6000000',
   'x-ms-client-request-id',
-  '7f5fa6ec-9b21-4d45-b4ad-e8c8dfaf95a5',
+  'dd29b7a8-4035-4bb3-991c-bf4473598816',
   'x-ms-version',
   '2019-02-02',
   'X-Content-Type-Options',
@@ -31,13 +31,13 @@ nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":tru
   'DataServiceId',
   "https://fakeaccount.table.core.windows.net/tableClientTestSASConnectionStringnode(PartitionKey='P4_node',RowKey='R4')",
   'Date',
-  'Wed, 09 Jun 2021 16:21:45 GMT'
+  'Tue, 07 Sep 2021 16:24:06 GMT'
 ]);
 
 nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":true})
   .get(`/tableClientTestSASConnectionStringnode(PartitionKey='P4_node',RowKey='R4')`)
   .query(true)
-  .reply(200, {"odata.metadata":"https://fakeaccount.table.core.windows.net/$metadata#tableClientTestSASConnectionStringnode/@Element","odata.etag":"W/\"datetime'2021-06-09T16%3A21%3A45.3613343Z'\"","PartitionKey":"P4_node","RowKey":"R4","Timestamp":"2021-06-09T16:21:45.3613343Z","testField@odata.type":"Edm.Int64","testField":"12345543221"}, [
+  .reply(200, {"odata.metadata":"https://fakeaccount.table.core.windows.net/$metadata#tableClientTestSASConnectionStringnode/@Element","odata.etag":"W/\"datetime'2021-09-07T16%3A24%3A07.3017257Z'\"","PartitionKey":"P4_node","RowKey":"R4","Timestamp":"2021-09-07T16:24:07.3017257Z","testField@odata.type":"Edm.Int64","testField":"12345543221"}, [
   'Cache-Control',
   'no-cache',
   'Transfer-Encoding',
@@ -45,13 +45,13 @@ nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":tru
   'Content-Type',
   'application/json;odata=minimalmetadata;streaming=true;charset=utf-8',
   'ETag',
-  `W/"datetime'2021-06-09T16%3A21%3A45.3613343Z'"`,
+  `W/"datetime'2021-09-07T16%3A24%3A07.3017257Z'"`,
   'Server',
   'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '654ef0d3-9002-001e-7c4b-5d5b52000000',
+  '00765408-e002-0038-3d04-a4c0e6000000',
   'x-ms-client-request-id',
-  '0a7a3617-1e62-4e5a-87db-7c36ba5db762',
+  '24b7ba37-7533-44cf-8894-36e5d0ae9cb8',
   'x-ms-version',
   '2019-02-02',
   'X-Content-Type-Options',
@@ -61,7 +61,7 @@ nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":tru
   'Access-Control-Allow-Origin',
   '*',
   'Date',
-  'Wed, 09 Jun 2021 16:21:45 GMT'
+  'Tue, 07 Sep 2021 16:24:07 GMT'
 ]);
 
 nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":true})
@@ -75,13 +75,13 @@ nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":tru
   'Server',
   'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '654ef0d8-9002-001e-804b-5d5b52000000',
+  '00765424-e002-0038-5604-a4c0e6000000',
   'x-ms-client-request-id',
-  '396db5b6-b76e-4146-8689-18fd1f93d9b1',
+  '4a76dac8-c7e2-4693-872e-fefae1a18888',
   'x-ms-version',
   '2019-02-02',
   'X-Content-Type-Options',
   'nosniff',
   'Date',
-  'Wed, 09 Jun 2021 16:21:45 GMT'
+  'Tue, 07 Sep 2021 16:24:07 GMT'
 ]);